Ver 5

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -1,12 +1,13 @@
 package ru.yandex.practicum.filmorate.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
 @Data
 @RequiredArgsConstructor
 public class Genre {
-   final private Integer id;
-   final private String name;
+    final private Long id;
+    final private String name;
 
 }
